move tests to test_structured_config
